Fetch library and frontend moved to their own files
